refactor(routing): move the parameters mapper out of the url generator
ParametersMapper moves to its own Routing\Mapper namespace, and PathGenerator is
dropped: the path is already carried by the operation metadata, so generating it
a second time at url generation was duplicated logic.
