+++ config/services/routing.php
@@ -5,16 +5,14 @@ declare(strict_types=1);
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use LAG\AdminBundle\Routing\Loader\ResourceRoutingLoader;
+use LAG\AdminBundle\Routing\Mapper\ParametersMapper;
+use LAG\AdminBundle\Routing\Mapper\ParametersMapperInterface;
 use LAG\AdminBundle\Routing\Route\RouteNameGenerator;
 use LAG\AdminBundle\Routing\Route\RouteNameGeneratorInterface;
 use LAG\AdminBundle\Routing\UrlGenerator\LinkUrlGenerator;
 use LAG\AdminBundle\Routing\UrlGenerator\LinkUrlGeneratorInterface;
 use LAG\AdminBundle\Routing\UrlGenerator\OperationUrlGenerator;
 use LAG\AdminBundle\Routing\UrlGenerator\OperationUrlGeneratorInterface;
-use LAG\AdminBundle\Routing\UrlGenerator\ParametersMapper;
-use LAG\AdminBundle\Routing\UrlGenerator\ParametersMapperInterface;
-use LAG\AdminBundle\Routing\UrlGenerator\PathGenerator;
-use LAG\AdminBundle\Routing\UrlGenerator\PathGeneratorInterface;
 use LAG\AdminBundle\Routing\UrlGenerator\UrlGenerator;
 use LAG\AdminBundle\Routing\UrlGenerator\UrlGeneratorInterface;
 
@@ -24,7 +22,6 @@ return static function (ContainerConfigurator $container): void {
     $services->set(ResourceRoutingLoader::class)
         ->args([
             '$requestParameter' => param('lag_admin.request_parameter'),
-            '$pathGenerator' => service(PathGeneratorInterface::class),
             '$resourceCollectionFactory' => service('lag_admin.resource.collection_factory'),
         ])
         ->tag('routing.loader')
@@ -52,7 +49,6 @@ return static function (ContainerConfigurator $container): void {
         ->alias('lag_admin.routing.link_url_generator', LinkUrlGeneratorInterface::class)
     ;
 
-    $services->set(PathGeneratorInterface::class, PathGenerator::class);
     $services->set(UrlGeneratorInterface::class, UrlGenerator::class)
         ->args([
             '$router' => service('router'),
