+++ tests/unit/Routing/UrlGenerator/ParametersMapperTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace LAG\AdminBundle\Tests\Unit\Routing\UrlGenerator;
 
-use LAG\AdminBundle\Routing\UrlGenerator\ParametersMapper;
+use LAG\AdminBundle\Routing\Mapper\ParametersMapper;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
@@ -17,18 +17,22 @@ final class ParametersMapperTest extends TestCase
     #[DataProvider('parameters')]
     public function itMapParameters(mixed $data, array $routeParameters, array $expectedParameters): void
     {
-        $parameters = $this->parametersMapper->map($data, $routeParameters);
+        $parameters = $this->parametersMapper->mapObjectToRouteParameters($data, $routeParameters);
 
         self::assertEquals($expectedParameters, $parameters);
     }
 
     public static function parameters(): iterable
     {
-        $stdClass = new \stdClass();
+        $obj = new \stdClass();
+        $obj->myProp = 'hello';
 
-        yield [$stdClass, [], []];
+        yield [$obj, [], []];
         yield [null, ['my_param' => 'my_value'], []];
         yield [['my_param' => 'my_value'], ['my_param' => null], ['my_param' => 'my_value']];
+        yield [$obj, ['myProp' => 'myProp'], ['myProp' => 'hello']];
+        yield [$obj, ['myProp'], ['myProp' => 'hello']];
+        yield [['key' => 'value'], ['key'], ['key' => 'value']];
     }
 
     protected function setUp(): void
